res merged trunk r27891:27983 into rendermanager branch.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/rendermanager@27985 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/bsptree.h
@@ -55,10 +55,10 @@ private:
   csArray<int> splitters;
 
   size_t FindBestSplitter (csTriangle* triangles, csPlane3* planes,
-	int num_triangles, const csVector3* vertices,
+	size_t num_triangles, const csVector3* vertices,
 	const csArray<int>& triidx);
   void Build (csTriangle* triangles, csPlane3* planes,
-	int num_triangles, const csVector3* vertices,
+	size_t num_triangles, const csVector3* vertices,
 	const csArray<int>& triidx);
   void Back2Front (const csVector3& pos, csDirtyAccessArray<int>& arr,
   	csSet<int>& used_indices);
@@ -75,7 +75,7 @@ public:
   /**
    * Build the BSP tree given the set of triangles.
    */
-  void Build (csTriangle* triangles, int num_triangles,
+  void Build (csTriangle* triangles, size_t num_triangles,
     const csVector3* vertices);
   void Build (CS::TriangleIndicesStream<int>& triangles, const csVector3* vertices);
   //@}
